+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - closes [#30](https://github.com/splashbyte/animated_toggle_switch/issues/30)
 - minor improvements
 - BREAKING: changes `separatorBuilder` parameters
+- BREAKING: moves all cursor parameters to `cursors`
 
 ## 0.8.0-beta.0 (2023-07-29)
 - adds tests for all `AnimatedToggleSwitch` constructors